+++ src/demo/models/index.js
@@ -1,3 +1,3 @@
-export {init as initTraining} from './training';
-export {init as initPredicting} from './predicting';
+export {init as initTraining} from './train';
+export {init as initPredicting} from './predict';
 export {init as initPond} from './pond';
