+++ src/demo/models/train.js
@@ -1,0 +1,9 @@
+import {setState, getState} from '../state';
+import {render} from '../renderer';
+
+export function init() {
+  // set state
+
+  // pass canvas as param?
+  render();
+}
