--- src/demo/models/training.js
@@ -1,1 +1,0 @@
-export function init() {}
